+++ odoo/local-src/connector_esb/tests/test_export_promotion_alcyon.py
@@ -14,6 +14,7 @@ class ExportPromotionAlcyonTestCase(ESBXMLTestCase):
         self.maxDiff = None
         self.timestamp = self.env.ref(
                 'connector_esb.esb_timestamp_promotion_alcyon')
+        self.timestamp.writer = 'local'
 
     @property
     def model(self):
@@ -28,6 +29,10 @@ class ExportPromotionAlcyonTestCase(ESBXMLTestCase):
         self.p2 = self.env['product.product'].create({
             'name': 'Unittest P2',
             'default_code': '0002'
+        })
+        self.p3 = self.env['product.product'].create({
+            'name': 'Unittest P3',
+            'default_code': '0003'
         })
         # Could be changed when export-product is merged ?
         self.ali = self.env['product.price.category'].create({
@@ -49,10 +54,17 @@ class ExportPromotionAlcyonTestCase(ESBXMLTestCase):
                     'price_category_id': self.ali.id,
                 }),
                 (0, False, {
-                    'applied_on': '2b_product_price_category',
+                    'applied_on': '3_global',
                     'product_id': self.p2.id,
                     'compute_price': 'percentage',
                     'percent_price': 8,
+                    'price_category_id': '',
+                }),
+                (0, False, {
+                    'applied_on': '2b_product_price_category',
+                    'product_id': self.p3.id,
+                    'compute_price': 'percentage',
+                    'percent_price': 0,
                     'price_category_id': self.alg.id,
                 }),
             ],
@@ -61,24 +73,50 @@ class ExportPromotionAlcyonTestCase(ESBXMLTestCase):
     def test_filename(self):
         self.check_filename('AlcyonPromotion_{0}.xml')
 
-    def test_mapper(self):
-        """ Testing mapper without id client """
+    def test_mapper_applied_on_2b(self):
+        """ Testing mapper on 2b_product_price_category """
         expected = {
             'AlcyonGroupId': 'Ref123',
             'Percent1': '5.00',
             'Percent2': '0',
             'ProductType': 'ALI'
         }
-        self.timestamp.writer = 'local'
         with self.backend.work_on(self.model._name,
                                   timestamp=self.timestamp) as work:
             mapper = work.component(usage='export.mapper')
-            rec = self.discount_pricelist_1.item_ids[0]
+            rec = self.discount_pricelist_1.item_ids.filtered(
+                    lambda r: r.product_id.id == self.p1.id)[0]
             self.assertDictEqual(mapper.map_record(rec).values(), expected)
+
+    def test_mapper_applied_on_3(self):
+        """ Testing mapper on 3_global """
+        expected = {
+            'AlcyonGroupId': 'Ref123',
+            'Percent1': '8.00',
+            'Percent2': '0',
+            'ProductType': ' ',
+        }
+        with self.backend.work_on(self.model._name,
+                                  timestamp=self.timestamp) as work:
+            mapper = work.component(usage='export.mapper')
+            rec = self.discount_pricelist_1.item_ids.filtered(
+                    lambda r: r.applied_on == '3_global')[0]
+            self.assertDictEqual(mapper.map_record(rec).values(), expected)
+
+    def test_getitems(self):
+        """If both percent are 0 do not include entry
+
+        And the percent 2 is always zero by the way
+        """
+        with self.backend.work_on(self.model._name,
+                                  timestamp=self.timestamp) as work:
+            exporter = work.component(usage='record.exporter.cron')
+            items = exporter.get_items(None)
+            self.assertEqual(len(items),
+                             len(self.discount_pricelist_1.item_ids) - 1)
 
     def test_export(self):
         """ Make a full export check with existing xml file"""
-        self.timestamp.writer = 'local'
         with self.backend.work_on(self.model._name,
                                   timestamp=self.timestamp) as work:
             exporter = work.component(usage='record.exporter.cron')